requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
opencv-python>=4.8
scikit-image>=0.22
scipy>=1.11
scikit-learn>=1.3
joblib>=1.3
PyYAML>=6.0
